feat: Create the GoalsContainer and CtaCard components, import the abstract logo SVG.
The About Us page holds a second version of a CTA, which will be a card to be shown in the end of the page and will problably take the users to the prices page. So, I decided to already create the component and submit the code in the next commits.