feat: Create the code for the PricingContainer component, refactor the home page.tsx.
While I was creating the page.tsx for the pricing page, I had the idea to create a shared component that would hold the whole pricing idea, and would be able to be shared between both home and pricing page.tsx components. So, I created a new PricesContainer component that would have the option between monthly and yearly prices and the available features in both options. That would make more sense than write the same code twice in different places, given the fact that they hold the same style and information.