refactor: Remove the framer motion usage and the dynamic use inside the MapTestimonials and TestimonialsCard components.
After thinking a little bit more, I found the challenge is a little bit harder than I expected. So I'm gonna think a new way to dynamically use the testimonials inside other pages, giving the fact that some components could be shared but I still did not find a way to do it. Will work on this later on.