feat: Create the laptop version of the courses page.
To keep the constancy of the project, I decided to create another script for the courses, but this time with the whole info, not only the ones from the previously created script. Maybe I can use both inside the CoursesFullContainer and save a few code lines in the coursesFullData script. Will think about that later. Also created the laptop version of the PageHero component, everything seems to be running smoothly. Will continue the project's responsive design.