feat: Create a ColOrganizer component to organize the number of columns outside the HomeSection component.
When I began the open course page, I realized that, if I wasn't using the HomeSection component, I'd have to repeat the code over and over to make a dynamic column grid, so, I decided to create a new component called ColOrganizer, that has all the logic involved in dynamically creating the grid area, with the desired number of columns. By doing this, I don't repeat the code and make it more readable and clean.